+++ aikau/src/main/resources/alfresco/layout/AlfTabContainer.js
@@ -177,13 +177,14 @@ define(["dojo/_base/declare",
         "alfresco/core/topics",
         "dijit/layout/TabContainer",
         "dijit/layout/ContentPane",
+        "dojo/Deferred",
         "dojo/dom-construct",
         "dojo/dom-class",
         "dojo/_base/lang",
         "dojo/_base/array",
         "jquery"], 
         function(declare, _WidgetBase, _TemplatedMixin, template, AlfCore, CoreWidgetProcessing, ResizeMixin, 
-                 topics, TabContainer, ContentPane, domConstruct, domClass, lang, array, $) {
+                 topics, TabContainer, ContentPane, Deferred, domConstruct, domClass, lang, array, $) {
    
    return declare([_WidgetBase, _TemplatedMixin, AlfCore, CoreWidgetProcessing, ResizeMixin], {
       
@@ -241,6 +242,17 @@ define(["dojo/_base/declare",
       doLayout: false,
 
       /**
+       * This can be configured to give all tabs some padding around their content. The amount of
+       * padding is controlled by the LESS variable "tab-padding"
+       * 
+       * @instance
+       * @type {boolean}
+       * @default
+       * @since 1.0.79
+       */
+      padded: false,
+
+      /**
        * This array is used to store widgets for delayed processing
        * 
        * @type {array}
@@ -294,6 +306,16 @@ define(["dojo/_base/declare",
        * @default
        */
       delayProcessingDefault: true,
+
+      /**
+       * A promise of the children of the [tabContainerWidget]{@link module:alfresco/layout/AlfTabContainer#tabContainerWidget}
+       * that is resolved when it is created.
+       * 
+       * @instance
+       * @type {object}
+       * @since 1.0.79
+       */
+      _tabContainerChildrenPromise: null,
 
       /**
        * Creates the "dijit/layout/TabContainer" wrapped by this widget and sets up associated
@@ -372,10 +394,14 @@ define(["dojo/_base/declare",
                {
                   this.alfSubscribe(this.tabDeletionTopic, lang.hitch(this, this.onTabDelete));
                }
-
                
                this.alfPublishResizeEvent(this.domNode);
                domClass.add(this.domNode, "alfresco-layout-AlfTabContainer--tabsDisplayed");
+
+               if (this._tabContainerChildrenPromise)
+               {
+                  this._tabContainerChildrenPromise.resolve(this.tabContainerWidget.getChildren());
+               }
             }
             catch(e)
             {
@@ -385,14 +411,32 @@ define(["dojo/_base/declare",
       },
 
       /**
+       * Returns a promise of the children of the 
+       * [tabContainerWidget]{@link module:alfresco/layout/AlfTabContainer#tabContainerWidget}
+       * that will be resolved when it is created.
+       * 
+       * @instance
+       * @return {promise} A promise of the children of the tab container
+       * @since 1.0.79
+       */
+      getTabContainerChildren: function alfresco_layout_AlfTabContainer__getTabContainerChildren() {
+         return this._tabContainerChildrenPromise;
+      },
+
+      /**
        * Calls [createTabContainer]{@link module:alfresco/layout/AlfTabContainer#createTabContainer}
        * to create the "dijit/layout/TabContainer" that is wrapped by this widget.
        * 
        * @instance
        */
       postCreate: function alfresco_layout_AlfTabContainer__postCreate() {
+         this._tabContainerChildrenPromise = new Deferred();
          this.createTabContainer();
          this.alfSetupResizeSubscriptions(this.onResize, this);
+         if (this.padded)
+         {
+            domClass.add(this.domNode, "alfresco-layout-AlfTabContainer--padded");
+         }
       },
 
       /**