+++ aikau/src/test/resources/testApp/WEB-INF/classes/alfresco/site-webscripts/alfresco/layout/AlfTabContainer.get.js
@@ -24,6 +24,7 @@ model.jsonModel = {
                            id: "TC",
                            name: "alfresco/layout/AlfTabContainer",
                            config: {
+                              padded: true,
                               tabSelectionTopic: "TABCONTAINER_SELECT_TAB_TOPIC",
                               tabDisablementTopic: "TABCONTAINER_DISABLE_TAB_TOPIC",
                               tabAdditionTopic: "TABCONTAINER_ADD_TAB_TOPIC",